+++ src/app/request.js
@@ -1,6 +1,20 @@
 import fetch from 'dva/fetch';
 import { Toast } from 'antd-mobile';
 import backToHome from '../utils/backToHome';
+
+let onLine = window.navigator.onLine;
+
+window.addEventListener('online', () => {
+	onLine = true;
+	Toast.info('网络恢复', 2);
+}, false);
+
+window.addEventListener('offline', () => {
+	onLine = false;
+	Toast.fail('网络断开', 2);
+}, false);
+
+
 function parseJSON(response) {
 	return response.json();
 }
@@ -29,7 +43,7 @@ function validation(responseJson) {
 			Toast.info('连接超时，即将返回主页', 2, () => {
 				backToHome();
 			});
-			throw responseJson.errmsg;
+			return { data: null, err: '连接超时，即将返回主页'};
 		default:
 			Toast.info(responseJson.errmsg, 2);
 			throw responseJson.errmsg;
@@ -41,6 +55,12 @@ function exception(err) {
 	return { data: null, err };
 }
 
+function offLine() {
+  return new Promise(resolve => {
+    resolve({ data: null, err: '无网络连接' });
+  });
+}
+
 
 /**
  * 发送网络请求，如果是POST请求需要传递bodyObject
@@ -48,6 +68,7 @@ function exception(err) {
  * @param {object} bodyObject 
  */
 export default function request(url, bodyObject) {
+	if(!onLine) return offLine();
 	let options = { credentials: 'include' };
 	if (bodyObject) options = {
 		method: 'POST',